Added the foundation with md generation from files.
Added all of the code to make graphQl search for .md files and
automatically transform them into Gatsby.js pages.

Created the new Template for each post.

Added the appropriate packages to generate the HTML from the Markdown as
well as the filesystem package to search inside the project for the .md
files of the posts.